pyproject.toml
[project]
name = "08-daypilot-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.133.0",
    "openai-agents>=0.10.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.41.0",
]
